Refactoring updated project structure to differentiate model, service and dao layers in text editor as well.
