+++ src/main/java/com/codecool/am_i_tea/dao/ProjectDAO.java
@@ -1,4 +1,6 @@
-package com.codecool.am_i_tea;
+package com.codecool.am_i_tea.dao;
+
+import com.codecool.am_i_tea.model.Project;
 
 public class ProjectDAO {
