+++ src/main/java/com/codecool/am_i_tea/dao/TextFileDAO.java
@@ -1,4 +1,6 @@
-package com.codecool.am_i_tea;
+package com.codecool.am_i_tea.dao;
+
+import com.codecool.am_i_tea.model.TextFile;
 
 public class TextFileDAO {
