+++ src/main/java/com/codecool/am_i_tea/model/TextFile.java
@@ -1,4 +1,4 @@
-package com.codecool.am_i_tea;
+package com.codecool.am_i_tea.model;
 
 public class TextFile {
